+++ src/main/java/com/bgmi/squad/bgmi/bean/Player.java
@@ -27,6 +27,6 @@ public class Player {
     private String firstName;
     private String lastName;
     private int age;
-    private int bgmiId;
+    private long bgmiId;
 
 }
